Split PROAdaptiveFC.cxx into per-concern translation units
Mechanical split of the 3.4k-line file into:
  - PROAdaptiveFC.cxx      mode dispatcher (run_adaptive_fc) only
  - PROAdaptiveFCmesh.cxx  throw generation, Wilks AMR prepass, meta-mesh
  - PROAdaptiveFCbank.cxx  boost serialisation + save/load, PE worker and
                           scheduler, asimov observables, classification
  - PROAdaptiveFCplot.cxx  all ROOT PDF/artifact output
  - PROAdaptiveFCinternal.h  cross-TU declarations in namespace PROfit::afc
                           (kept in src/ — internal, not public API)

Function bodies are moved verbatim; formerly file-local functions that now
cross TU boundaries move from internal linkage into PROfit::afc, everything
else stays static in its new TU. All boost archive I/O remains in the bank
TU. Deleted the unreferenced struct CellState and its stale banner (its
planned sequential-stopping use was superseded by the deterministic
doubling rule in schedule_pes). No behavior change.
