+++ src/PROAdaptiveFCinternal.h
@@ -1,0 +1,217 @@
+/**
+ * @file PROAdaptiveFCinternal.h
+ * @brief Internal cross-TU declarations for the adaptive Feldman-Cousins pipeline.
+ * @author PROfit Collaboration
+ *
+ * @internal Not installed and not part of the public API — that is
+ * inc/PROAdaptiveFC.h. The adaptive-FC implementation is split across
+ * src/PROAdaptiveFC.cxx (mode dispatcher), src/PROAdaptiveFCmesh.cxx
+ * (throws, Wilks prepass, meta-mesh), src/PROAdaptiveFCbank.cxx (PE bank
+ * serialisation, PE worker/scheduler, asimov observables, classification)
+ * and src/PROAdaptiveFCplot.cxx (all ROOT PDF/artifact output). Everything
+ * declared here lives in namespace PROfit::afc so formerly file-local names
+ * stay out of the library-wide PROfit namespace.
+ */
+#ifndef PROADAPTIVEFCINTERNAL_H
+#define PROADAPTIVEFCINTERNAL_H
+
+#include "PROAdaptiveFC.h"
+#include "PROmodel.h"
+
+#include <Eigen/Eigen>
+
+#include <cstdint>
+#include <string>
+#include <vector>
+
+namespace PROfit {
+namespace afc {
+
+// --------------------------------------------------------------------
+//  Throw generation + meta-mesh (src/PROAdaptiveFCmesh.cxx)
+// --------------------------------------------------------------------
+
+std::vector<PROmesh::AMRResult> generate_throws(
+    const PROconfig &config,
+    const PROpeller &prop,
+    const PROsyst   &systs,
+    const PROmodel  &model,
+    const PROfitterConfig &fitconfig,
+    PROseed &proseed,
+    const Eigen::VectorXf &fakeDataParams,
+    const AdaptiveFCConfig &acfg,
+    int nthreads,
+    size_t xaxis_idx, size_t yaxis_idx,
+    MultiPROgressBar &progress);
+
+MetaMesh build_meta_mesh(const std::vector<PROmesh::AMRResult> &throws,
+                         float p_thresh,
+                         int baseline_level);
+
+void compute_cell_centers(const MetaMesh &mm,
+                          bool xlog, bool ylog,
+                          std::vector<float> &cx_out,
+                          std::vector<float> &cy_out);
+
+PROdata generate_pseudo_experiment_data(
+    const PROconfig &config,
+    const PROpeller &prop,
+    const PROsyst   &systs,
+    const PROmodel  &model,
+    const Eigen::VectorXf &fakeDataParams,
+    bool binned,
+    PROseed &proseed);
+
+// --------------------------------------------------------------------
+//  PE bank, asimov observables, classification (src/PROAdaptiveFCbank.cxx)
+// --------------------------------------------------------------------
+
+void schedule_pes(const AdaptiveFCConfig &acfg,
+                  const PROconfig &config,
+                  const PROpeller &prop,
+                  const PROsyst   &systs,
+                  const PROmodel  &model,
+                  const PROfitterConfig &fitconfig,
+                  PROseed &proseed,
+                  const Eigen::MatrixXf &L,
+                  size_t xaxis_idx, size_t yaxis_idx,
+                  const std::vector<float> &cell_x_model,
+                  const std::vector<float> &cell_y_model,
+                  PEBank &bank_out,
+                  int nthreads,
+                  MultiPROgressBar &progress,
+                  int progress_bar_idx,
+                  AdaptiveFCResult &result_out);
+
+// Observed Δχ² on the asimov dataset at every meta-mesh cell.
+//
+// The global (syst+osc) fit is computed once — physics floats freely, so the
+// result is μ-independent. Per-cell fits then re-pin the two scanned axes at
+// each cell center and run a syst-only fit (splines float; non-scanned
+// physics params pinned at model->default_val(i)).
+struct AsimovObs {
+    float chi2_osc_global = 0.0f;        ///< Single global best fit.
+    std::vector<float> chi2_syst;        ///< Per-cell syst-only chi^2.
+    std::vector<float> dchi2_obs;        ///< chi2_syst[c] - chi2_osc_global per cell.
+};
+
+AsimovObs compute_asimov_obs(
+    const PROconfig &config,
+    const PROpeller &prop,
+    const PROsyst   &systs,
+    const PROmodel  &model,
+    const PROfitterConfig &fitconfig,
+    const PROdata   &asimov_data,
+    const std::string &chi2_kind,
+    bool binned,
+    size_t xaxis_idx, size_t yaxis_idx,
+    const std::vector<float> &cell_x_model,
+    const std::vector<float> &cell_y_model,
+    PROseed &proseed,
+    int nthreads,
+    MultiPROgressBar &progress,
+    int bar_idx);
+
+// Per-cell, per-CL verdict map produced by classify_against_bank.
+struct CellVerdict {
+    float crit_dchi2 = 0.0f;  ///< Empirical critical Δχ² at the bank's α-quantile.
+    bool  included  = false;  ///< dchi2_obs ≤ crit_dchi2 at this CL.
+    bool  decidable = false;  ///< Bank had enough PEs to give a stable quantile.
+};
+
+std::vector<std::vector<CellVerdict>> classify_against_bank(
+    const PEBank &bank,
+    const AsimovObs &obs,
+    const std::vector<float> &cl_targets,
+    int min_pes_for_decision);
+
+// --------------------------------------------------------------------
+//  ROOT PDF / artifact output (src/PROAdaptiveFCplot.cxx)
+// --------------------------------------------------------------------
+
+void write_slice1_diagnostics(
+    const std::vector<PROmesh::AMRResult> &throws,
+    const MetaMesh &mm,
+    const PROmodel &model,
+    const PROsyst  &systs,
+    const AdaptiveFCConfig &acfg,
+    size_t xaxis_idx, size_t yaxis_idx,
+    AdaptiveFCResult &result_out);
+
+void plot_pebank_summary_pdf(const PEBank &bank,
+                             const std::string &filename,
+                             const std::string &bank_path,
+                             const std::string &xlabel,
+                             const std::string &ylabel,
+                             bool logx, bool logy,
+                             bool xlog_axis, bool ylog_axis);
+
+void plot_pebank_pes_multipage_pdf(const PEBank &bank,
+                                   const std::string &filename,
+                                   const std::string &xlabel,
+                                   const std::string &ylabel,
+                                   bool xlog_axis, bool ylog_axis);
+
+void plot_asimov_verdict_pdf(
+    const PEBank &bank,
+    const AsimovObs &obs,
+    const std::vector<float> &cl_targets,
+    const std::vector<std::vector<CellVerdict>> &verdicts,
+    const std::string &filename,
+    const std::string &bank_path,
+    const std::string &xlabel,
+    const std::string &ylabel,
+    bool logx, bool logy,
+    bool xlog_axis, bool ylog_axis);
+
+void plot_asimov_contour_pdf(
+    const PEBank &bank,
+    const AsimovObs &obs,
+    const std::vector<float> &cl_targets,
+    const std::vector<std::vector<CellVerdict>> &verdicts,
+    const std::string &filename,
+    const std::string &xlabel,
+    const std::string &ylabel,
+    bool logx, bool logy,
+    bool xlog_axis, bool ylog_axis,
+    bool draw_truth_marker,
+    float truth_x_phys,
+    float truth_y_phys);
+
+void save_asimov_root(
+    const PEBank &bank,
+    const AsimovObs &obs,
+    const std::vector<float> &cl_targets,
+    const std::vector<std::vector<CellVerdict>> &verdicts,
+    const std::string &filename,
+    bool xlog_axis, bool ylog_axis);
+
+void plot_brazil_band_pdf(
+    const PEBank &bank,
+    const std::vector<std::vector<float>> &inclusion_frac, // [cl_idx][cell_idx]
+    const std::vector<float> &cl_targets,
+    const std::string &filename,
+    const std::string &bank_path,
+    const std::string &xlabel,
+    const std::string &ylabel,
+    bool logx, bool logy,
+    bool xlog_axis, bool ylog_axis,
+    bool draw_truth_marker,
+    float truth_x_phys,
+    float truth_y_phys,
+    int n_throws);
+
+void save_brazil_root(
+    const PEBank &bank,
+    const std::vector<std::vector<std::vector<uint8_t>>> &per_throw_verdicts, // [t][cl][cell]
+    const std::vector<std::vector<float>> &per_throw_dchi2,                   // [t][cell]
+    const std::vector<float> &per_throw_global_chi2,                          // [t]
+    const std::vector<std::vector<float>> &inclusion_frac,                    // [cl][cell]
+    const std::vector<float> &cl_targets,
+    const std::string &filename,
+    bool xlog_axis, bool ylog_axis);
+
+}  // namespace afc
+}  // namespace PROfit
+
+#endif
